Remove bundler bin files
These files where most likely added via the `$ bundler gem ...` command,
and assumed the common pattern of using the `exe/` dir for executables,
and the `bin/` dir for developer scripts (not to be included with the
gem).

Since the gemspec uses `git ls-files` to generate the files for the
packaged gem, there is a clean way of rejecting files, and since these
files really don't make much sense for this gem, deleting them is
simpler.