Remove `git` dependency from gemspec file
Does a few things:

* Allows the memory_analyzer.gemspec to be evaluated without requiring
  `git` to be present on the system
* Removes unwanted files from being packaged in the gem (.travis.yml,
  .gitignore, Rakefile, Gemfile, etc.)

Minor disadvantage is that if a file is added to the root directory of
the project that is required to be packaged with the gem, this has to be
added manually to the `spec.files` line, but that shouldn't be a big
deal, or terribly common (this might be a CHANGELOG file at best from
what is currently here) as most files should be added to lib).

The LICENSE.txt and README.md were chosen as items that might be useful
to a user perusing the installed packaged gem, but these subject for
discussion/debate.