Support gzipped heap dump files
When comparing sizes of a memory dump file when compressed and not:

    $ ls -lh tmp/sample_dumps
    total 1256264
    -rw-r--r--  1 user  staff   576M Jan 01 19:08 dump
    -rw-r--r--  1 user  staff    37M Jan 01 19:09 dump.gz

The disk space requirements to keep around a gzipped file are far more
beneficial instead of it's uncompressed counterpart, and the time
difference and memory requirements to parse a gzip file versus a
uncompressed file are next to none.

    $ time ruby bin/memory_analyzer tmp/sample_dumps/dump.gz
    Parsing: |========================================| Time: 00:03:11
    Mem: 3695.37890625MiB

    real    3m14.405s
    user    3m11.839s
    sys     0m2.357s
    $ time ruby bin/memory_analyzer tmp/sample_dumps/dump
    Parsing: |========================================| Time: 00:03:11
    Mem: 3625.39453125MiB

    real    3m13.220s
    user    3m10.516s
    sys     0m2.522s

Adds the `#io_class` which will look at the filename's extension to
determine if it is a gzipped file (probably a better way to do this),
and then uses chooses the proper IO class to read the file with.  The
`ProgressBar` `:total` option also makes use of this method.

If the file is a gzip file, it will require the zlib library and use the
`Zlib::GzipReader` class to parse the file.

The `#parse_file` method needed to be modified slightly to use a common
API between the `Zlib::GzipReader` and `File` classes, but changes to
the method functionally equivalent to the old implementation.